Got SubComments working. Added UI for them. Move models to model.py class. Added memcache (from Tim James example)
