+++ templates/_base.html
@@ -9,7 +9,6 @@
 	<script src="https://ajax.googleapis.com/ajax/libs/jquery/1.11.3/jquery.min.js"></script>
 	<!-- Latest compiled JavaScript -->
 	<script src="http://maxcdn.bootstrapcdn.com/bootstrap/3.3.5/js/bootstrap.min.js"></script>
-	<script src="http://ajax.googleapis.com/ajax/libs/jquery/1.5.2/jquery.min.js"></script>
 	<script src="/static/script0001.js"></script>
 	<script src="/static/maps.js"></script>
 	<link rel="stylesheet" type="text/css" href="/static/style.css">
@@ -68,10 +67,6 @@
 				<div class="modal-body" style="text-align:justify">
 					<p><em>Pitt Yak</em> is the best application known to man.</p>
 				</div>
-				
-				<div class="modal-footer">
-					<button type="button" class="btn btn-default" data-dismiss="modal">Close</button>
-				</div>
 			</div>
 		</div>
 	</div>
@@ -110,7 +105,6 @@
 				<div class="modal-footer">
 					<button type="submit" class="btn btn-default">Submit</button>
 					</form>
-					<button type="button" class="btn btn-default" data-dismiss="modal">Close</button>
 				</div>
 			</div>
 		</div>
